+++ src/decorators/module.ts
@@ -2,7 +2,7 @@ export function Module(options: Object | string)
 export function Module(target: Object, key: string)
 export function Module(first: any, second?: string) {
   //Bare decorator (no params)
-  console.log("options", first, second);
+  // console.log("options", first, second);
   if (second) propDecorator(null)(first, second);
   //Decorator with params
   else return propDecorator(first);
@@ -14,7 +14,7 @@ function propDecorator(options?: any) {
       if (!options['controllers'] && typeof options === 'object') {
         throw Error("controllers parameter 'Module' must be an valid object");
       }
-    } else throw Error("Fist parameter 'Module' must be a string or an object");
+    } else throw Error("Fist parameter 'Module' must be an object");
 
     delete target[key]; // Todo: validate this line
     return options['controllers'];