+++ src/server.ts
@@ -5,8 +5,6 @@ import { IServer } from "./interfaces/IServer";
 Vue.use(Servue)
 
 export function Server(controllers: any[] | any): IServer & vuejs.Vue {
-  console.log("Controllers:", controllers)
-
   var server: any = new Vue({
     mixins: controllers
   });
@@ -15,7 +13,7 @@ export function Server(controllers: any[] | any): IServer & vuejs.Vue {
   // server.info = "David"
   // const port = 3000;
   // server.serve('localhost', port, function () {
-  // console.log(`Listening ${port}...`)
+  ///// console.log(`Listening ${port}...`)
   // });
   // server.getList( { params: { send: 1323 } });
 }